Created spa and van frontend

+++ flask_api/main.py
@@ -1,5 +1,6 @@
 import flask
 from flask import request, jsonify
+from flask import json
 from flask.helpers import flash
 import flash_controller
 import blog_controller
@@ -44,7 +45,8 @@ def all_books():
         })
         response_object['message'] = 'Book added!'
     else:
-        response_object['books'] = BOOKS
+        books = book_controller.get_books()
+        response_object['books'] = books
     return jsonify(response_object)
 
 @app.route('/books/<book_id>', methods=['PUT', 'DELETE'])
@@ -74,42 +76,47 @@ def remove_book(book_id):
 
 # Blog Post Routes ***********************************************************
 
-@app.route("/blog/posts", methods=["GET"])
-def get_posts():
-    posts = blog_controller.get_posts()
-    return jsonify(posts)
+@app.route('/blog/posts', methods=['GET', 'POST'])
+def all_posts():
+    response_object = {'status': 'success'}
+    if request.method == 'POST':
+        post_data = request.get_json()
+        post_date = post_data["post_date"]
+        edit_date = post_data["edit_date"]
+        location = post_data["location"] 
+        author = post_data["author"]
+        author_id = post_data["author_id"]
+        post_title = post_data["post_title"] 
+        post_subtitle = post_data["post_subtitle"]
+        post_body = post_data["post_body"]
+        post_tags = post_data["post_tags"]
+        result = blog_controller.insert_post(post_date, edit_date, location, author, author_id, post_title, post_subtitle, post_body, post_tags)
+        response_object['message'] = 'Post Added!'
+    else:
+        posts = blog_controller.get_posts()
+        response_object['posts'] = posts
+    return jsonify(response_object)
 
-@app.route("/blog/post", methods=["POST"])
-def insert_post():
-    post_info = request.get_json()
-    post_date = post_info["post_date"]
-    edit_date = post_info["edit_date"]
-    location = post_info["location"] 
-    author = post_info["author"]
-    author_id = post_info["author_id"]
-    post_title = post_info["post_title"] 
-    post_body = post_info["post_body"]
-    result = blog_controller.insert_post(post_date, edit_date, location, author, author_id, post_title, post_body)
-    return jsonify(result)
-
-@app.route("/blog/post", methods=["PUT"])
-def update_post():
-    post_info = request.get_json()
-    post_id = post_info["post_id"]
-    post_date = post_info["post_date"]
-    edit_date = post_info["edit_date"]
-    location = post_info["location"] 
-    author = post_info["author"]
-    author_id = post_info["author_id"]
-    post_title = post_info["post_title"] 
-    post_body = post_info["post_body"]
-    result = blog_controller.update_post(post_id, post_date, edit_date, location, author, author_id, post_title, post_body)
-    return jsonify(result)
-
-@app.route("/blog/post/<id>", methods=["DELETE"])
-def delete_post(id):
-    result = blog_controller.delete_post(id)
-    return jsonify(result)
+@app.route("/blog/post/<id>", methods=['PUT', 'DELETE'])
+def single_post(id):
+    response_object = {'status': 'success'}
+    if request.method == 'PUT':
+        post_data = request.get_json()
+        post_date = post_data["post_date"]
+        edit_date = post_data["edit_date"]
+        location = post_data["location"] 
+        author = post_data["author"]
+        author_id = post_data["author_id"]
+        post_title = post_data["post_title"] 
+        post_subtitle = post_data["post_subtitle"]
+        post_body = post_data["post_body"]
+        post_tags = post_data["post_tags"]
+        blog_controller.update_post(post_date, edit_date, location, author, author_id, post_title, post_subtitle, post_body, post_tags)
+        response_object['message'] = 'Post Updated!'
+    if request.method == 'DELETE':
+        blog_controller.delete_post(id)
+        response_object['message'] = 'Post Removed'
+    return jsonify(response_object)
 
 @app.route("/blog/post/<id>", methods=["GET"])
 def get_post_by_id(id):
